requirements.txt
fastapi==0.115.5
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.9.0
PyJWT==2.10.0
PyJWT==2.10.0
python-dotenv==1.0.1
python_jose==3.3.0
sqlmodel==0.0.22
uvicorn==0.32.0
